fix: Clean path only for files, not directory
